requirements.txt
pandas
matplotlib
seaborn
numpy
pathlib
datasets
openai
python-dotenv 
tqdm
scikit-learn
tiktoken